requirements.txt
getrpimodel
smbus2
boto3
